refactor(nocodb): svs -> svc
