+++ packages/nocodb/src/lib/services/index.ts
@@ -27,7 +27,7 @@ export * as attachmentService from './attachment.svc';
 export * as hookFilterService from './hookFilter.svc';
 export * as dataService from './dbData';
 export * as bulkDataService from './dbData/bulkData';
-export * as dataAliasNestedService from './dbData/dataAliasNested.svs';
+export * as dataAliasNestedService from './dbData/dataAliasNested.svc';
 export * as cacheService from './cache.svc';
 export * as auditService from './audit.svc';
 export * as swaggerService from './apiDocs';
@@ -35,4 +35,4 @@ export * as userService from './user';
 export * as syncService from './sync';
 export * from './public';
 export * as orgTokenService from './orgToken.svc';
-export * as orgTokenServiceEE from './ee/orgToken.svs';
+export * as orgTokenServiceEE from './ee/orgToken.svc';
