+++ packages/nocodb/src/lib/services/public/index.ts
@@ -1,5 +1,5 @@
-import * as publicDataService from './publicData.svs';
-import * as publicDataExportApis from './publicDataExport.svs';
-import * as publicMetaService from './publicMeta.svs';
+import * as publicDataService from './publicData.svc';
+import * as publicDataExportApis from './publicDataExport.svc';
+import * as publicMetaService from './publicMeta.svc';
 
 export { publicDataService, publicDataExportApis, publicMetaService };
